@@ -11,10 +11,11 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/volatiletech/sqlboiler/v4/importers"
+	"github.com/aarondl/sqlboiler/v4/importers"
+	"github.com/aarondl/strmangle"
 
-	"github.com/volatiletech/sqlboiler/v4/drivers"
-	"github.com/volatiletech/sqlboiler/v4/drivers/mocks"
+	"github.com/aarondl/sqlboiler/v4/drivers"
+	"github.com/aarondl/sqlboiler/v4/drivers/mocks"
 )
 
 var state *State
@@ -80,7 +81,7 @@ func testNew(t *testing.T, aliases Aliases) {
 		PkgName:    "models",
 		OutFolder:  out,
 		NoTests:    true,
-		DriverConfig: map[string]interface{}{
+		DriverConfig: map[string]any{
 			drivers.ConfigSchema:    "schema",
 			drivers.ConfigBlacklist: []string{"hangars"},
 		},
@@ -106,7 +107,7 @@ func testNew(t *testing.T, aliases Aliases) {
 		t.Fatalf("go env GOMOD cmd execution failed: %s", err)
 	}
 
-	cmd = exec.Command("go", "mod", "init", "github.com/volatiletech/sqlboiler-test")
+	cmd = exec.Command("go", "mod", "init", "github.com/aarondl/sqlboiler-test")
 	cmd.Dir = state.Config.OutFolder
 	cmd.Stderr = buf
 
@@ -116,7 +117,7 @@ func testNew(t *testing.T, aliases Aliases) {
 		fmt.Println()
 	}
 
-	cmd = exec.Command("go", "mod", "edit", fmt.Sprintf("-replace=github.com/volatiletech/sqlboiler/v4=%s", filepath.Dir(string(goModFilePath))))
+	cmd = exec.Command("go", "mod", "edit", fmt.Sprintf("-replace=github.com/aarondl/sqlboiler/v4=%s", filepath.Dir(string(goModFilePath))))
 	cmd.Dir = state.Config.OutFolder
 	cmd.Stderr = buf
 
@@ -226,6 +227,7 @@ func TestProcessTypeReplacements(t *testing.T) {
 	s := new(State)
 	s.Config = &Config{}
 	s.Config.Imports.BasedOnType = make(map[string]importers.Set)
+	s.Config.SkipReplacedEnumTypes = true
 	domainStr := "a_domain"
 	s.Tables = []drivers.Table{
 		{
@@ -347,5 +349,12 @@ func TestProcessTypeReplacements(t *testing.T) {
 	}
 	if i := s.Config.Imports.BasedOnType["excellent.NamedType"].ThirdParty[0]; i != `"rock.com/excellent-name"` {
 		t.Error("imports were not adjusted")
+	}
+
+	expectedDisallows := []string{
+		"int", "null.String", // Must NOT contain replaced "excellent.Type".
+	}
+	if disallowList := s.Config.DiscardedEnumTypes; len(strmangle.SetComplement(disallowList, expectedDisallows)) != 0 {
+		t.Error("expected disallows:", disallowList)
 	}
 }
