@@ -10,12 +10,12 @@ import (
 func TestSetFromInterface(t *testing.T) {
 	t.Parallel()
 
-	setIntf := map[string]interface{}{
-		"standard": []interface{}{
+	setIntf := map[string]any{
+		"standard": []any{
 			"hello",
 			"there",
 		},
-		"third_party": []interface{}{
+		"third_party": []any{
 			"there",
 			"hello",
 		},
@@ -43,13 +43,13 @@ func TestSetFromInterface(t *testing.T) {
 func TestMapFromInterface(t *testing.T) {
 	t.Parallel()
 
-	mapIntf := map[string]interface{}{
-		"test_main": map[string]interface{}{
-			"standard": []interface{}{
+	mapIntf := map[string]any{
+		"test_main": map[string]any{
+			"standard": []any{
 				"hello",
 				"there",
 			},
-			"third_party": []interface{}{
+			"third_party": []any{
 				"there",
 				"hello",
 			},
@@ -83,14 +83,14 @@ func TestMapFromInterface(t *testing.T) {
 func TestMapFromInterfaceAltSyntax(t *testing.T) {
 	t.Parallel()
 
-	mapIntf := []interface{}{
-		map[string]interface{}{
+	mapIntf := []any{
+		map[string]any{
 			"name": "test_main",
-			"standard": []interface{}{
+			"standard": []any{
 				"hello",
 				"there",
 			},
-			"third_party": []interface{}{
+			"third_party": []any{
 				"there",
 				"hello",
 			},
@@ -175,7 +175,7 @@ func TestAddTypeImports(t *testing.T) {
 			`"fmt"`,
 		},
 		ThirdParty: List{
-			`"github.com/volatiletech/sqlboiler/v4/boil"`,
+			`"github.com/aarondl/sqlboiler/v4/boil"`,
 		},
 	}
 
@@ -186,8 +186,8 @@ func TestAddTypeImports(t *testing.T) {
 			`"time"`,
 		},
 		ThirdParty: List{
-			`"github.com/volatiletech/null/v8"`,
-			`"github.com/volatiletech/sqlboiler/v4/boil"`,
+			`"github.com/aarondl/null/v8"`,
+			`"github.com/aarondl/sqlboiler/v4/boil"`,
 		},
 	}
 
@@ -200,7 +200,7 @@ func TestAddTypeImports(t *testing.T) {
 	imps := NewDefaultImports()
 
 	imps.BasedOnType = Map{
-		"null.Time": Set{ThirdParty: List{`"github.com/volatiletech/null/v8"`}},
+		"null.Time": Set{ThirdParty: List{`"github.com/aarondl/null/v8"`}},
 		"time.Time": Set{Standard: List{`"time"`}},
 	}
 
@@ -217,8 +217,8 @@ func TestAddTypeImports(t *testing.T) {
 			`"time"`,
 		},
 		ThirdParty: List{
-			`"github.com/volatiletech/null/v8"`,
-			`"github.com/volatiletech/sqlboiler/v4/boil"`,
+			`"github.com/aarondl/null/v8"`,
+			`"github.com/aarondl/sqlboiler/v4/boil"`,
 		},
 	}
 
@@ -234,11 +234,11 @@ func TestMergeSet(t *testing.T) {
 
 	a := Set{
 		Standard:   List{"fmt"},
-		ThirdParty: List{"github.com/volatiletech/sqlboiler/v4", "github.com/volatiletech/null/v8"},
+		ThirdParty: List{"github.com/aarondl/sqlboiler/v4", "github.com/aarondl/null/v8"},
 	}
 	b := Set{
 		Standard:   List{"os"},
-		ThirdParty: List{"github.com/volatiletech/sqlboiler/v4"},
+		ThirdParty: List{"github.com/aarondl/sqlboiler/v4"},
 	}
 
 	c := mergeSet(a, b)
@@ -246,8 +246,8 @@ func TestMergeSet(t *testing.T) {
 	if c.Standard[0] != "fmt" && c.Standard[1] != "os" {
 		t.Errorf("Wanted: fmt, os got: %#v", c.Standard)
 	}
-	if c.ThirdParty[0] != "github.com/volatiletech/null/v8" && c.ThirdParty[1] != "github.com/volatiletech/sqlboiler/v4" {
-		t.Errorf("Wanted: github.com/volatiletech/sqlboiler, github.com/volatiletech/null/v8 got: %#v", c.ThirdParty)
+	if c.ThirdParty[0] != "github.com/aarondl/null/v8" && c.ThirdParty[1] != "github.com/aarondl/sqlboiler/v4" {
+		t.Errorf("Wanted: github.com/aarondl/sqlboiler, github.com/aarondl/null/v8 got: %#v", c.ThirdParty)
 	}
 }
 
