@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"regexp"
 
-	"github.com/volatiletech/sqlboiler/v4/boil"
-	"github.com/volatiletech/sqlboiler/v4/drivers"
+	"github.com/aarondl/sqlboiler/v4/boil"
+	"github.com/aarondl/sqlboiler/v4/drivers"
 )
 
 // joinKind is the type of join
@@ -31,7 +31,7 @@ type Query struct {
 	loadMods map[string]Applicator
 
 	delete     bool
-	update     map[string]interface{}
+	update     map[string]any
 	withs      []argClause
 	selectCols []string
 	count      bool
@@ -74,33 +74,33 @@ type where struct {
 
 	clause      string
 	orSeparator bool
-	args        []interface{}
+	args        []any
 }
 
 type in struct {
 	clause      string
 	orSeparator bool
-	args        []interface{}
+	args        []any
 }
 
 type argClause struct {
 	clause string
-	args   []interface{}
+	args   []any
 }
 
 type rawSQL struct {
 	sql  string
-	args []interface{}
+	args []any
 }
 
 type join struct {
 	kind   joinKind
 	clause string
-	args   []interface{}
+	args   []any
 }
 
 // Raw makes a raw query, usually for use with bind
-func Raw(query string, args ...interface{}) *Query {
+func Raw(query string, args ...any) *Query {
 	return &Query{
 		rawSQL: rawSQL{
 			sql:  query,
@@ -110,7 +110,7 @@ func Raw(query string, args ...interface{}) *Query {
 }
 
 // RawG makes a raw query using the global boil.Executor, usually for use with bind
-func RawG(query string, args ...interface{}) *Query {
+func RawG(query string, args ...any) *Query {
 	return Raw(query, args...)
 }
 
@@ -122,6 +122,12 @@ func (q *Query) Exec(exec boil.Executor) (sql.Result, error) {
 		fmt.Fprintln(boil.DebugWriter, args)
 	}
 	return exec.Exec(qs, args...)
+}
+
+// QueryRowG executes the query for the One finisher and returns a row.
+// It uses the global executer.
+func (q *Query) QueryRowG() *sql.Row {
+	return q.QueryRow(boil.GetDB())
 }
 
 // QueryRow executes the query for the One finisher and returns a row
@@ -155,6 +161,12 @@ func (q *Query) ExecContext(ctx context.Context, exec boil.ContextExecutor) (sql
 	return exec.ExecContext(ctx, qs, args...)
 }
 
+// QueryRowContextG executes the query for the One finisher and returns a row.
+// It uses the global executer.
+func (q *Query) QueryRowContextG(ctx context.Context) *sql.Row {
+	return q.QueryRowContext(ctx, boil.GetContextDB())
+}
+
 // QueryRowContext executes the query for the One finisher and returns a row
 func (q *Query) QueryRowContext(ctx context.Context, exec boil.ContextExecutor) *sql.Row {
 	qs, args := BuildQuery(q)
@@ -177,6 +189,12 @@ func (q *Query) QueryContext(ctx context.Context, exec boil.ContextExecutor) (*s
 	return exec.QueryContext(ctx, qs, args...)
 }
 
+// ExecG executes a query that does not need a row returned.
+// It uses the global executer.
+func (q *Query) ExecG() (sql.Result, error) {
+	return q.Exec(boil.GetDB())
+}
+
 // ExecP executes a query that does not need a row returned
 // It will panic on error
 func (q *Query) ExecP(exec boil.Executor) sql.Result {
@@ -186,6 +204,18 @@ func (q *Query) ExecP(exec boil.Executor) sql.Result {
 	}
 
 	return res
+}
+
+// ExecGP executes a query that does not need a row returned.
+// It uses the global executer and panics on error.
+func (q *Query) ExecGP() sql.Result {
+	return q.ExecP(boil.GetDB())
+}
+
+// QueryG executes the query for the All finisher and returns multiple rows.
+// It uses the global executer.
+func (q *Query) QueryG() (*sql.Rows, error) {
+	return q.Query(boil.GetDB())
 }
 
 // QueryP executes the query for the All finisher and returns multiple rows
@@ -199,13 +229,65 @@ func (q *Query) QueryP(exec boil.Executor) *sql.Rows {
 	return rows
 }
 
+// QueryGP executes the query for the All finisher and returns multiple rows.
+// It uses the global executer and panics on error.
+func (q *Query) QueryGP() *sql.Rows {
+	return q.QueryP(boil.GetDB())
+}
+
+// ExecContextG executes a query that does not need a row returned with a given context.
+// It uses the global executer.
+func (q *Query) ExecContextG(ctx context.Context) (sql.Result, error) {
+	return q.ExecContext(ctx, boil.GetContextDB())
+}
+
+// ExecContextP executes a query with a given context that does not need a row returned.
+// It panics on error.
+func (q *Query) ExecContextP(ctx context.Context, exec boil.ContextExecutor) sql.Result {
+	result, err := q.ExecContext(ctx, exec)
+	if err != nil {
+		panic(boil.WrapErr(err))
+	}
+
+	return result
+}
+
+// ExecContextGP executes a query with a given context that does not need a row returned.
+// It uses the global executer and panics on error.
+func (q *Query) ExecContextGP(ctx context.Context) sql.Result {
+	return q.ExecContextP(ctx, boil.GetContextDB())
+}
+
+// QueryContextG executes the query for the All finisher with a given context and returns multiple rows.
+// It uses the global executer.
+func (q *Query) QueryContextG(ctx context.Context) (*sql.Rows, error) {
+	return q.QueryContext(ctx, boil.GetContextDB())
+}
+
+// QueryContextP executes the query for the All finisher with a given context and returns multiple rows.
+// It panics on error.
+func (q *Query) QueryContextP(ctx context.Context, exec boil.ContextExecutor) *sql.Rows {
+	rows, err := q.QueryContext(ctx, exec)
+	if err != nil {
+		panic(boil.WrapErr(err))
+	}
+
+	return rows
+}
+
+// QueryContextGP executes the query for the All finisher with a given context and returns multiple rows.
+// It uses the global executer and panics on error.
+func (q *Query) QueryContextGP(ctx context.Context) *sql.Rows {
+	return q.QueryContextP(ctx, boil.GetContextDB())
+}
+
 // SetDialect on the query.
 func SetDialect(q *Query, dialect *drivers.Dialect) {
 	q.dialect = dialect
 }
 
 // SetSQL on the query.
-func SetSQL(q *Query, sql string, args ...interface{}) {
+func SetSQL(q *Query, sql string, args ...any) {
 	q.rawSQL = rawSQL{sql: sql, args: args}
 }
 
@@ -213,7 +295,7 @@ func SetSQL(q *Query, sql string, args ...interface{}) {
 // query text does not need to be re-generated, useful
 // if you're performing the same query with different arguments
 // over and over.
-func SetArgs(q *Query, args ...interface{}) {
+func SetArgs(q *Query, args ...any) {
 	q.rawSQL.args = args
 }
 
@@ -282,7 +364,7 @@ func SetComment(q *Query, comment string) {
 }
 
 // SetUpdate on the query.
-func SetUpdate(q *Query, cols map[string]interface{}) {
+func SetUpdate(q *Query, cols map[string]any) {
 	q.update = cols
 }
 
@@ -350,22 +432,22 @@ func AppendFullOuterJoin(q *Query, clause string, args ...interface{}) {
 }
 
 // AppendHaving on the query.
-func AppendHaving(q *Query, clause string, args ...interface{}) {
+func AppendHaving(q *Query, clause string, args ...any) {
 	q.having = append(q.having, argClause{clause: clause, args: args})
 }
 
 // AppendWhere on the query.
-func AppendWhere(q *Query, clause string, args ...interface{}) {
+func AppendWhere(q *Query, clause string, args ...any) {
 	q.where = append(q.where, where{clause: clause, args: args})
 }
 
 // AppendIn on the query.
-func AppendIn(q *Query, clause string, args ...interface{}) {
+func AppendIn(q *Query, clause string, args ...any) {
 	q.where = append(q.where, where{kind: whereKindIn, clause: clause, args: args})
 }
 
 // AppendNotIn on the query.
-func AppendNotIn(q *Query, clause string, args ...interface{}) {
+func AppendNotIn(q *Query, clause string, args ...any) {
 	q.where = append(q.where, where{kind: whereKindNotIn, clause: clause, args: args})
 }
 
@@ -421,12 +503,12 @@ func AppendGroupBy(q *Query, clause string) {
 }
 
 // AppendOrderBy on the query.
-func AppendOrderBy(q *Query, clause string, args ...interface{}) {
+func AppendOrderBy(q *Query, clause string, args ...any) {
 	q.orderBy = append(q.orderBy, argClause{clause: clause, args: args})
 }
 
 // AppendWith on the query.
-func AppendWith(q *Query, clause string, args ...interface{}) {
+func AppendWith(q *Query, clause string, args ...any) {
 	q.withs = append(q.withs, argClause{clause: clause, args: args})
 }
 
