@@ -101,7 +101,7 @@ func TestAppendWhere(t *testing.T) {
 		t.Errorf("args wrong: %#v", q.where)
 	}
 
-	q.where = []where{{clause: expect, args: []interface{}{5, 3}}}
+	q.where = []where{{clause: expect, args: []any{5, 3}}}
 	if q.where[0].clause != expect {
 		t.Errorf("Expected %s, got %v", expect, q.where)
 	}
@@ -176,7 +176,7 @@ func TestAppendIn(t *testing.T) {
 		t.Errorf("args wrong: %#v", q.where)
 	}
 
-	q.where = []where{{clause: expect, args: []interface{}{5, 3}}}
+	q.where = []where{{clause: expect, args: []any{5, 3}}}
 	if q.where[0].clause != expect {
 		t.Errorf("Expected %s, got %v", expect, q.where)
 	}
@@ -262,7 +262,7 @@ func TestAppendOrderBy(t *testing.T) {
 	}
 
 	q.orderBy = []argClause{
-		{"col1 desc, col2 asc", []interface{}{}},
+		{"col1 desc, col2 asc", []any{}},
 	}
 	if len(q.orderBy) != 1 && q.orderBy[0].clause != expect {
 		t.Errorf("Expected %s, got %s", expect, q.orderBy[0].clause)
@@ -289,7 +289,7 @@ func TestAppendHaving(t *testing.T) {
 		t.Errorf("Expected %v, got %v %v", 10, q.having[0].args[0], q.having[1].args[0])
 	}
 
-	q.having = []argClause{{clause: expect, args: []interface{}{10}}}
+	q.having = []argClause{{clause: expect, args: []any{10}}}
 	if len(q.having) != 1 && (q.having[0].clause != expect || q.having[0].args[0] != 10) {
 		t.Errorf("Expected %s, got %s %v", expect, q.having[0], q.having[0].args[0])
 	}
@@ -350,7 +350,7 @@ func TestSetUpdate(t *testing.T) {
 	t.Parallel()
 
 	q := &Query{}
-	SetUpdate(q, map[string]interface{}{"test": 5})
+	SetUpdate(q, map[string]any{"test": 5})
 
 	if q.update["test"] != 5 {
 		t.Errorf("Wrong update, got %v", q.update)
@@ -371,7 +371,7 @@ func TestSetDelete(t *testing.T) {
 func TestSetArgs(t *testing.T) {
 	t.Parallel()
 
-	args := []interface{}{2}
+	args := []any{2}
 	q := &Query{rawSQL: rawSQL{}}
 	SetArgs(q, args...)
 
@@ -459,7 +459,7 @@ func TestAppendInnerJoin(t *testing.T) {
 
 	q.joins = []join{{kind: JoinInner,
 		clause: "thing=$1 AND stuff=$2",
-		args:   []interface{}{2, 5},
+		args:   []any{2, 5},
 	}}
 
 	if len(q.joins) != 1 {
@@ -502,7 +502,7 @@ func TestAppendLeftOuterJoin(t *testing.T) {
 
 	q.joins = []join{{kind: JoinOuterLeft,
 		clause: "thing=$1 AND stuff=$2",
-		args:   []interface{}{2, 5},
+		args:   []any{2, 5},
 	}}
 
 	if len(q.joins) != 1 {
@@ -545,7 +545,7 @@ func TestAppendRightOuterJoin(t *testing.T) {
 
 	q.joins = []join{{kind: JoinOuterRight,
 		clause: "thing=$1 AND stuff=$2",
-		args:   []interface{}{2, 5},
+		args:   []any{2, 5},
 	}}
 
 	if len(q.joins) != 1 {
@@ -588,7 +588,7 @@ func TestAppendFullOuterJoin(t *testing.T) {
 
 	q.joins = []join{{kind: JoinOuterFull,
 		clause: "thing=$1 AND stuff=$2",
-		args:   []interface{}{2, 5},
+		args:   []any{2, 5},
 	}}
 
 	if len(q.joins) != 1 {
@@ -631,7 +631,7 @@ func TestAppendWith(t *testing.T) {
 
 	q.withs = []argClause{{
 		clause: "other_cte AS (SELECT * FROM other_table WHERE thing=$1 AND stuff=$2)",
-		args:   []interface{}{3, 7},
+		args:   []any{3, 7},
 	}}
 
 	if len(q.withs) != 1 {
