@@ -14,8 +14,8 @@ import (
 	"unicode"
 
 	"github.com/friendsofgo/errors"
-	"github.com/volatiletech/sqlboiler/v4/boil"
-	"github.com/volatiletech/strmangle"
+	"github.com/aarondl/sqlboiler/v4/boil"
+	"github.com/aarondl/strmangle"
 )
 
 // Identifies what kind of object we're binding to
@@ -38,7 +38,7 @@ const (
 // result into the passed in object pointer.
 // It panics on error.
 // Also see documentation for Bind() and Query.Bind()
-func (q *Query) BindP(ctx context.Context, exec boil.Executor, obj interface{}) {
+func (q *Query) BindP(ctx context.Context, exec boil.Executor, obj any) {
 	if err := q.Bind(ctx, exec, obj); err != nil {
 		panic(boil.WrapErr(err))
 	}
@@ -48,8 +48,15 @@ func (q *Query) BindP(ctx context.Context, exec boil.Executor, obj interface{}) 
 // the result into the passed in object pointer.
 // It uses the global executor.
 // Also see documentation for Bind() and Query.Bind()
-func (q *Query) BindG(ctx context.Context, obj interface{}) error {
+func (q *Query) BindG(ctx context.Context, obj any) error {
 	return q.Bind(ctx, boil.GetDB(), obj)
+}
+
+// BindGP executes the query and inserts the result into the passed in object pointer.
+// It uses the global executer and panics on error.
+// Also see documentation for Bind() and Query.Bind()
+func (q *Query) BindGP(ctx context.Context, obj any) {
+	q.BindP(ctx, boil.GetDB(), obj)
 }
 
 // Bind inserts the rows into the passed in object pointer, because the caller
@@ -107,13 +114,13 @@ func (q *Query) BindG(ctx context.Context, obj interface{}) error {
 //
 // For custom objects that want to use eager loading, please see the
 // loadRelationships function.
-func Bind(rows *sql.Rows, obj interface{}) error {
-	structType, sliceType, singular, err := bindChecks(obj)
+func Bind(rows *sql.Rows, obj any) error {
+	structType, _, singular, err := bindChecks(obj)
 	if err != nil {
 		return err
 	}
 
-	return bind(rows, obj, structType, sliceType, singular)
+	return bind(rows, obj, structType, singular)
 }
 
 // Bind executes the query and inserts the
@@ -125,8 +132,8 @@ func Bind(rows *sql.Rows, obj interface{}) error {
 // be using load* methods that support context as the first parameter.
 //
 // Also see documentation for Bind()
-func (q *Query) Bind(ctx context.Context, exec boil.Executor, obj interface{}) error {
-	structType, sliceType, bkind, err := bindChecks(obj)
+func (q *Query) Bind(ctx context.Context, exec boil.Executor, obj any) error {
+	structType, _, bkind, err := bindChecks(obj)
 	if err != nil {
 		return err
 	}
@@ -140,7 +147,7 @@ func (q *Query) Bind(ctx context.Context, exec boil.Executor, obj interface{}) e
 	if err != nil {
 		return errors.Wrap(err, "bind failed to execute query")
 	}
-	if err = bind(rows, obj, structType, sliceType, bkind); err != nil {
+	if err = bind(rows, obj, structType, bkind); err != nil {
 		if innerErr := rows.Close(); innerErr != nil {
 			return errors.Wrapf(err, "error on rows.Close after bind error: %+v", innerErr)
 		}
@@ -163,7 +170,7 @@ func (q *Query) Bind(ctx context.Context, exec boil.Executor, obj interface{}) e
 
 // bindChecks resolves information about the bind target, and errors if it's not an object
 // we can bind to.
-func bindChecks(obj interface{}) (structType reflect.Type, sliceType reflect.Type, bkind bindKind, err error) {
+func bindChecks(obj any) (structType reflect.Type, sliceType reflect.Type, bkind bindKind, err error) {
 	typ := reflect.TypeOf(obj)
 	kind := typ.Kind()
 
@@ -216,7 +223,7 @@ func bindChecks(obj interface{}) (structType reflect.Type, sliceType reflect.Typ
 	}
 }
 
-func bind(rows *sql.Rows, obj interface{}, structType, sliceType reflect.Type, bkind bindKind) error {
+func bind(rows *sql.Rows, obj any, structType reflect.Type, bkind bindKind) error {
 	cols, err := rows.Columns()
 	if err != nil {
 		return errors.Wrap(err, "bind failed to get column names")
@@ -233,29 +240,22 @@ func bind(rows *sql.Rows, obj interface{}, structType, sliceType reflect.Type, b
 		return err
 	}
 
-	var oneStruct reflect.Value
-	if bkind == kindSliceStruct {
-		oneStruct = reflect.Indirect(reflect.New(structType))
-	}
-
 	foundOne := false
 Rows:
 	for rows.Next() {
 		foundOne = true
 		var newStruct reflect.Value
-		var pointers []interface{}
+		var pointers []any
 
 		switch bkind {
 		case kindStruct:
 			pointers = PtrsFromMapping(reflect.Indirect(reflect.ValueOf(obj)), mapping)
 		case kindSliceStruct:
-			pointers = PtrsFromMapping(oneStruct, mapping)
+			newStruct = reflect.Indirect(reflect.New(structType))
+			pointers = PtrsFromMapping(newStruct, mapping)
 		case kindPtrSliceStruct:
 			newStruct = makeStructPtr(structType)
 			pointers = PtrsFromMapping(reflect.Indirect(newStruct), mapping)
-		}
-		if err != nil {
-			return err
 		}
 
 		if err := rows.Scan(pointers...); err != nil {
@@ -265,9 +265,7 @@ Rows:
 		switch bkind {
 		case kindStruct:
 			break Rows
-		case kindSliceStruct:
-			ptrSlice.Set(reflect.Append(ptrSlice, oneStruct))
-		case kindPtrSliceStruct:
+		case kindSliceStruct, kindPtrSliceStruct:
 			ptrSlice.Set(reflect.Append(ptrSlice, newStruct))
 		}
 	}
@@ -329,8 +327,8 @@ ColLoop:
 
 // PtrsFromMapping expects to be passed an addressable struct and a mapping
 // of where to find things. It pulls the pointers out referred to by the mapping.
-func PtrsFromMapping(val reflect.Value, mapping []uint64) []interface{} {
-	ptrs := make([]interface{}, len(mapping))
+func PtrsFromMapping(val reflect.Value, mapping []uint64) []any {
+	ptrs := make([]any, len(mapping))
 	for i, m := range mapping {
 		ptrs[i] = ptrFromMapping(val, m, true).Interface()
 	}
@@ -339,8 +337,8 @@ func PtrsFromMapping(val reflect.Value, mapping []uint64) []interface{} {
 
 // ValuesFromMapping expects to be passed an addressable struct and a mapping
 // of where to find things. It pulls the pointers out referred to by the mapping.
-func ValuesFromMapping(val reflect.Value, mapping []uint64) []interface{} {
-	ptrs := make([]interface{}, len(mapping))
+func ValuesFromMapping(val reflect.Value, mapping []uint64) []any {
+	ptrs := make([]any, len(mapping))
 	for i, m := range mapping {
 		ptrs[i] = ptrFromMapping(val, m, false).Interface()
 	}
@@ -351,7 +349,7 @@ func ValuesFromMapping(val reflect.Value, mapping []uint64) []interface{} {
 // for things on.
 func ptrFromMapping(val reflect.Value, mapping uint64, addressOf bool) reflect.Value {
 	if mapping == 0 {
-		var ignored interface{}
+		var ignored any
 		return reflect.ValueOf(&ignored)
 	}
 	for i := 0; i < 8; i++ {
@@ -508,7 +506,7 @@ func (b *mappingCache) mapping(cols []string) ([]uint64, error) {
 //
 // Choosing not to use the DefaultParameterConverter here because sqlboiler doesn't generate
 // pointer columns.
-func Equal(a, b interface{}) bool {
+func Equal(a, b any) bool {
 	if (a == nil && b != nil) || (a != nil && b == nil) {
 		return false
 	}
@@ -570,7 +568,7 @@ func Equal(a, b interface{}) bool {
 }
 
 // isNumeric tests if i is a numeric value.
-func isNumeric(i interface{}) bool {
+func isNumeric(i any) bool {
 	switch i.(type) {
 	case int,
 		int8,
@@ -591,9 +589,9 @@ func isNumeric(i interface{}) bool {
 
 // parseNumeric tries to parse s as t.
 // t must be a numeric type.
-func parseNumeric(s string, t reflect.Type) interface{} {
+func parseNumeric(s string, t reflect.Type) any {
 	var (
-		res interface{}
+		res any
 		err error
 	)
 	switch t.Kind() {
@@ -620,7 +618,7 @@ func parseNumeric(s string, t reflect.Type) interface{} {
 
 // Assign assigns a value to another using reflection.
 // Dst must be a pointer.
-func Assign(dst, src interface{}) {
+func Assign(dst, src any) {
 	// Fast path for []byte since it's one of the
 	// most frequent other "ids" we'll be assigning.
 	if db, ok := dst.(*[]byte); ok {
@@ -669,7 +667,7 @@ func Assign(dst, src interface{}) {
 	}
 }
 
-func upgradeNumericTypes(i interface{}) interface{} {
+func upgradeNumericTypes(i any) any {
 	switch t := i.(type) {
 	case int:
 		return int64(t)
@@ -699,7 +697,7 @@ func upgradeNumericTypes(i interface{}) interface{} {
 // This whole function makes assumptions that whatever type
 // dst is, will be compatible with whatever came out of the Valuer.
 // We handle the types that driver.Value could possibly be.
-func assignValue(dst interface{}, val driver.Value) {
+func assignValue(dst any, val driver.Value) {
 	dstType := reflect.TypeOf(dst).Elem()
 	dstVal := reflect.ValueOf(dst).Elem()
 
@@ -758,7 +756,7 @@ func IsValuerNil(val driver.Valuer) bool {
 
 // IsNil is a more generic version of IsValuerNil, will check to make sure it's
 // not a valuer first.
-func IsNil(val interface{}) bool {
+func IsNil(val any) bool {
 	if val == nil {
 		return true
 	}
