@@ -62,7 +62,7 @@ func (d Decimal) Value() (driver.Value, error) {
 }
 
 // Scan implements sql.Scanner.
-func (d *Decimal) Scan(val interface{}) error {
+func (d *Decimal) Scan(val any) error {
 	newD, err := decimalScan(d.Big, val, false)
 	if err != nil {
 		return err
@@ -81,6 +81,24 @@ func (d *Decimal) UnmarshalJSON(data []byte) error {
 	return d.Big.UnmarshalJSON(data)
 }
 
+// MarshalText marshals a decimal value
+func (d Decimal) MarshalText() ([]byte, error) {
+	if d.Big == nil {
+		return nullBytes, nil
+	}
+
+	return d.Big.MarshalText()
+}
+
+// UnmarshalText allows marshalling text into a null pointer
+func (d *Decimal) UnmarshalText(data []byte) error {
+	if d.Big == nil {
+		d.Big = new(decimal.Big)
+	}
+
+	return d.Big.UnmarshalText(data)
+}
+
 // Randomize implements sqlboiler's randomize interface
 func (d *Decimal) Randomize(nextInt func() int64, fieldType string, shouldBeNull bool) {
 	d.Big = randomDecimal(nextInt, fieldType, false)
@@ -92,7 +110,7 @@ func (n NullDecimal) Value() (driver.Value, error) {
 }
 
 // Scan implements sql.Scanner.
-func (n *NullDecimal) Scan(val interface{}) error {
+func (n *NullDecimal) Scan(val any) error {
 	newD, err := decimalScan(n.Big, val, true)
 	if err != nil {
 		return err
@@ -116,6 +134,31 @@ func (n *NullDecimal) UnmarshalJSON(data []byte) error {
 	}
 
 	return n.Big.UnmarshalJSON(data)
+}
+
+// MarshalText marshals a decimal value
+func (n NullDecimal) MarshalText() ([]byte, error) {
+	if n.Big == nil {
+		return nullBytes, nil
+	}
+
+	return n.Big.MarshalText()
+}
+
+// UnmarshalText allows marshalling text into a null pointer
+func (n *NullDecimal) UnmarshalText(data []byte) error {
+	if bytes.Equal(data, nullBytes) {
+		if n != nil {
+			n.Big = nil
+		}
+		return nil
+	}
+
+	if n.Big == nil {
+		n.Big = decimal.WithContext(DecimalContext)
+	}
+
+	return n.Big.UnmarshalText(data)
 }
 
 // String impl
@@ -186,7 +229,7 @@ func decimalValue(d *decimal.Big, canNull bool) (driver.Value, error) {
 	return d.String(), nil
 }
 
-func decimalScan(d *decimal.Big, val interface{}, canNull bool) (*decimal.Big, error) {
+func decimalScan(d *decimal.Big, val any, canNull bool) (*decimal.Big, error) {
 	if val == nil {
 		if !canNull {
 			return nil, errors.New("null cannot be scanned into decimal")
