@@ -13,7 +13,7 @@ import (
 	"text/template"
 
 	"github.com/friendsofgo/errors"
-	"github.com/volatiletech/sqlboiler/v4/importers"
+	"github.com/aarondl/sqlboiler/v4/importers"
 )
 
 // Copied from the go source
@@ -25,7 +25,7 @@ var (
 )
 
 var (
-	noEditDisclaimerFmt = `// Code generated by SQLBoiler%s(https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+	noEditDisclaimerFmt = `// Code generated by SQLBoiler%s(https://github.com/aarondl/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 `
@@ -306,6 +306,8 @@ func getLongExt(filename string) string {
 }
 
 func getOutputFilename(tableName string, isTest, isGo bool) string {
+	tableName = strings.ReplaceAll(strings.ReplaceAll(tableName, `/`, `_`), `\`, `_`)
+
 	if strings.HasPrefix(tableName, "_") {
 		tableName = "und" + tableName
 	}
